% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{ensemble_member_generator}
\alias{ensemble_member_generator}
\title{Create an ensemble member generator}
\usage{
ensemble_member_generator(tstart, tend, tdiff_prng)
}
\arguments{
\item{tstart}{the starting time as \code{POSIXct}}

\item{tend}{the minimum end time as \code{POSIXct}}

\item{tdiff_prng}{the pseudo-random number generator for the time differences
between event beginnings.}
}
\value{
a numeric \code{vector} of simulated event beginning times
}
\description{
Create an ensemble member generator
}
