% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{ensemble_probabilities_times}
\alias{ensemble_probabilities_times}
\title{Get event occurrence probabilities from ensemble}
\usage{
ensemble_probabilities_times(ensemble, event_duration, times)
}
\arguments{
\item{ensemble}{the ensemble}

\item{event_duration}{the duration of an event}

\item{times}{the times to get the probabilities for}
}
\description{
Get event occurrence probabilities from ensemble
}
