% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{event_simulation_ensemble}
\alias{event_simulation_ensemble}
\title{Run an event occurrence simulation ensemble}
\usage{
event_simulation_ensemble(events, tend, size = 100,
  tdiff_prng = prng_from_sample(diff(as.numeric(events) -
  as.numeric(min(events)))))
}
\arguments{
\item{events}{the event occurrences as \code{POSIXct}}

\item{tend}{the minimum forecast end time to simulate as \code{POSIXct}}

\item{size}{the amount of ensemble members}

\item{tdiff_prng}{the pseudo-random number generator for the time differences
between event beginnings. Defaults to \code{prng_from_sample(events)}.}
}
\value{
a \code{list} of ensemble members
}
\description{
Run an event occurrence simulation ensemble
}
