% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prng_from_sample}
\alias{prng_from_sample}
\title{Get a PRNG from a sample}
\usage{
prng_from_sample(x, cdf = ecdf(x), xmin = min(x, na.rm = T), xmax = max(x,
  na.rm = T))
}
\arguments{
\item{x}{the sample, an numeric vector}

\item{cdf}{the cdf of the sample's distribution. If unspecified, the
empirical cdf is determined with \code{ecdf}.}

\item{xmin}{the lowest possible value of the sample's distribution.
Defaults to \code{min(x)}.}

\item{xmax}{the highest possible value of the sample's distribution.
Defaults to \code{max(x)}.}
}
\description{
\code{prng_from_sample} takes an arbitrary sample of data and returns a
pseudo-random number generator.
}
